Cleanup and refactor into separate files
